:recycle: refactor: 게시글 작성/수정 공통으로 쓰이는 로직 컴포넌트로 분리
